requirements.txt
python-telegram-bot
yt-dlp
ffmpeg
Flask
gunicorn
tdqm

